WIP Refactor slices
