+++ site/pages/our-work/case-study/case-study-components/style.css
@@ -48,22 +48,6 @@
   padding-bottom: 10px;
 }
 
-.caseStudyContainer {
-  max-width: 1100px;
-  justify-content: center;
-  margin: 0 auto;
-  padding-top: 30px;
-}
-
-.caseStudyContent {
-  display: flex;
-  max-width: containerMaxWidth;
-  width: 100%;
-  flex-direction: column-reverse;
-  margin-bottom: 20px;
-  padding-bottom: 20px;
-}
-
 @media mediumScreen {
   .section {
     padding-bottom: 48px;
@@ -77,24 +61,6 @@
   .quote {
     padding-left: 30px;
     padding-bottom: 0px;
-  }
-  .caseStudyContent {
-    display: flex;
-    margin: 0;
-    padding: 0;
-  }
-
-  .row {
-    flex-direction: row;
-  }
-
-  .row-reverse {
-    flex-direction: row-reverse;
-  }
-
-  .caseStudyContainer {
-    padding: 60px 0 60px;
-    border-bottom: 1px solid linesOnWhite;
   }
 }
