--- site/pages/our-work/slices/car-trawler/style.css
@@ -1,104 +1,0 @@
-@value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../../css/_colors.css";
-@value mediumScreen, largeScreen from "../../../../css/_sizes.css";
-
-.imageWrapper {
-  display: flex;
-  flex: 1;
-  justify-content: center;
-  align-items: center;
-}
-
-_:-ms-fullscreen,
-:root .imageWrapper {
-  min-height: 20vh;
-}
-
-.imageContainer {
-  min-width: 260px;
-  /* min-height: 205px; */
-  position: relative;
-}
-
-.imageWrapper img {
-  display: block;
-  width: 100%;
-  margin-top: 15px;
-}
-
-.caseStudyTextContainer {
-  padding: 0;
-  margin-top: 30px;
-}
-
-.caseStudyTextContainerHeader {
-  composes: fontM from '../../../../css/typography/_fonts.css';
-  composes: serif from '../../../../css/typography/_fonts.css';
-  color: badgerBlack;
-  margin-top: 20px;
-}
-
-.description {
-  composes: fontXS from '../../../../css/typography/_fonts.css';
-  composes: boldSansSerif from '../../../../css/typography/_fonts.css';
-  color: badgerBlack;
-  margin-top: 10px;
-}
-
-.readmore {
-  composes: boldSansSerif from '../../../../css/typography/_fonts.css';
-  border-bottom: 2px solid badgerRedOnWhite;
-  text-decoration: none;
-  color: badgerBlack;
-  display: inline-block;
-  margin-top: 30px;
-}
-
-.readmore:hover {
-  color: badgerRedOnWhite;
-}
-
-.links {
-  text-align: right;
-  border-bottom: 1px solid linesOnWhite;
-  padding-bottom: 30px;
-}
-
-@media mediumScreen {
-  .caseStudyTextContainer {
-    margin-top: 0;
-    padding: 0 30px;
-    flex-basis: 50%;
-  }
-
-  .imageWrapper {
-    flex-basis: 50%;
-  }
-
-  .imageContainer {
-    min-width: 400px;
-  }
-
-  .imageText {
-    font-size: 20px;
-  }
-
-  .caseStudyTextContainerHeader {
-    margin-top: 20px;
-  }
-
-  .description {
-    margin-top: 20px;
-  }
-
-  .links {
-    text-align: left;
-    border: none;
-    padding-bottom: 0;
-  }
-}
-
-@media largeScreen {
-  .caseStudyTextContainer {
-    padding-left: 90px;
-  }
-}
