+++ site/pages/our-work/slices/shared/styles.css
@@ -1,6 +1,22 @@
 @value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../../css/_colors.css";
 @value mediumScreen, largeScreen from "../../../../css/_sizes.css";
 
+.caseStudyContainer {
+  max-width: 1100px;
+  justify-content: center;
+  margin: 0 auto;
+  padding-top: 30px;
+}
+
+.caseStudyContent {
+  display: flex;
+  max-width: containerMaxWidth;
+  width: 100%;
+  flex-direction: column-reverse;
+  margin-bottom: 20px;
+  padding-bottom: 20px;
+}
+
 .imageWrapper {
   display: flex;
   flex: 1;
@@ -15,8 +31,6 @@ _:-ms-fullscreen,
 
 .imageContainer {
   min-width: 260px;
-  min-height: 205px;
-  background-color: #abc9c1;
   position: relative;
 }
 
@@ -26,32 +40,7 @@ _:-ms-fullscreen,
   margin-top: 15px;
 }
 
-.orders {
-  position: absolute;
-  right: 10px;
-  top: 5px;
-  text-align: right;
-}
-
-.visits {
-  position: absolute;
-  left: 10px;
-  bottom: 5px;
-}
-
-.imagePercentage {
-  composes: fontM from '../../../../css/typography/_fonts.css';
-  composes: serif from '../../../../css/typography/_fonts.css';
-  color: badgerBlack;
-}
-
-.imageText {
-  composes: fontS from '../../../../css/typography/_fonts.css';
-  composes: lightSerif from '../../../../css/typography/_fonts.css';
-  color: badgerBlack;
-  font-size: 16px;
-}
-
+.caseStudyTextContainerReverse,
 .caseStudyTextContainer {
   padding: 0;
   margin-top: 30px;
@@ -91,6 +80,26 @@ _:-ms-fullscreen,
 }
 
 @media mediumScreen {
+  .row {
+    flex-direction: row;
+  }
+
+  .row-reverse {
+    flex-direction: row-reverse;
+  }
+
+  .caseStudyContainer {
+    padding: 60px 0 60px;
+    border-bottom: 1px solid linesOnWhite;
+  }
+
+  .caseStudyContent {
+    display: flex;
+    margin: 0;
+    padding: 0;
+  }
+
+  .caseStudyTextContainerReverse,
   .caseStudyTextContainer {
     margin-top: 0;
     padding: 0 30px;
@@ -103,10 +112,6 @@ _:-ms-fullscreen,
 
   .imageContainer {
     min-width: 400px;
-  }
-
-  .imageText {
-    font-size: 20px;
   }
 
   .caseStudyTextContainerHeader {
@@ -122,26 +127,11 @@ _:-ms-fullscreen,
     border: none;
     padding-bottom: 0;
   }
-
-  .orders {
-    right: 10px;
-    top: 10px;
-  }
-
-  .visits {
-    left: 10px;
-    bottom: 10px;
-  }
 }
 
 @media largeScreen {
-  .orders {
-    right: 30px;
-    top: 30px;
-  }
-
-  .visits {
-    left: 30px;
-    bottom: 30px;
+  .caseStudyTextContainerReverse,
+  .caseStudyTextContainer {
+    padding-left: 90px;
   }
 }
